requirements.txt
torch>=2.1.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pymupdf>=1.23.0
python-docx>=0.8.11
numpy>=1.24.0
scikit-learn>=1.3.0
accelerate>=0.24.0
requests>=2.31.0
urllib3>=1.26.0